Refactor size hooks to correct directory
