+++ game/api/index.ts
@@ -1,5 +1,3 @@
 export { getResult } from "./getResult";
 export { sampleSize } from "./sampleSize";
 export { updateHints } from "./updateHints";
-export { useKeySizes } from "./useKeySizes";
-export { useTileSizes } from "./useTileSizes";
